+++ builder_pkgs/mockito/test/end2end/dart_js_interop_test.mocks.dart
@@ -4,6 +4,7 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
+
 import 'dart:js_interop' as _i1;
 
 import 'package:mockito/mockito.dart' as _i2;